APPs/BankingApp/requirements.txt
amazon-dax-client==2.0.3
antlr4-python3-runtime==4.9.3
bcrypt==3.1.7
boto3==1.26.1
botocore==1.29.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Session==0.3.1
fpdf==1.7.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==1.0.1
kiwisolver==1.1.0
lml==0.0.9
MarkupSafe==1.1.1
numpy==1.21.1
pycparser==2.19
pyecharts-jupyter-installer==0.0.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.22.0
s3transfer==0.6.0
six==1.14.0
texttable==1.6.2
urllib3==1.25.8
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
